Enable storing RPM dependencies
Each RPM can have arbitrary number of dependencies of multiple types.
The types are hard-coded, and can not be extended at run-time.

With this patch, it is possible to store and retrieve the dependencies.
There are tests for these parts.

Existing tests have been updated to pass: for RPM tests there are some
minor updates. The compose/package API and its variants are updated not
to include dependencies in their output.

All changes to dependencies are logged in change sets as single update
to the RPM.

There is a single filter for each dependency type. It allows to specify
an optional version. When a version is used in the filter, it removes
all packages that are incompatible with that version restriction. The
behaviour is documented with examples.

The version comparisons are implemented with a function adhering to PEP
0440 [0], which seems to be powerful enough to do everything required.
There is a wrapper that makes it work with epochs. It also works with
releases.

There is an additional single filter `has_no_deps`, which filters
packages that have some or don't have any dependency (this works across
all types).

[0]: https://www.python.org/dev/peps/pep-0440/

Changes in other parts of codebase:

 * Custom boolean filter is updated to allow removing duplicates from the
   response.

JIRA: PDC-955